Second Commit

+++ Illyria - The Last Defense/Assets/Scripts/FromCharacterToCharacterJson.cs
@@ -2,6 +2,6 @@
 {
     public static CharacterJson ConvertTo(Character c)
     {
-        return new CharacterJson{ Current_Level = c.Level_Current, Items = c.Items, Name = c.name, Stars = c.Stars};
+        return new CharacterJson{ Current_Level = c.Level_Current, Items = c.Items, Name = c.name, Stars = (int)c.Stars};
     }
 }
